+++ tests/test_metrics/test_metrics.py
@@ -47,8 +47,8 @@ from tests.utils import create_dummy_functional_metric
         (RMSE(), "RMSE(mode = 'per-segment', )"),
         (MedAE(), "MedAE(mode = 'per-segment', )"),
         (MSLE(), "MSLE(mode = 'per-segment', )"),
-        (MAPE(), "MAPE(mode = 'per-segment', )"),
-        (SMAPE(), "SMAPE(mode = 'per-segment', )"),
+        (MAPE(), "MAPE(mode = 'per-segment', missing_mode = 'error', )"),
+        (SMAPE(), "SMAPE(mode = 'per-segment', missing_mode = 'error', )"),
         (R2(), "R2(mode = 'per-segment', )"),
         (Sign(), "Sign(mode = 'per-segment', )"),
         (MaxDeviation(), "MaxDeviation(mode = 'per-segment', )"),
@@ -184,8 +184,8 @@ def test_invalid_nans_pred(metric_class, train_test_dfs):
         RMSE(),
         MedAE(),
         MSLE(),
-        MAPE(),
-        SMAPE(),
+        MAPE(missing_mode="error"),
+        SMAPE(missing_mode="error"),
         R2(),
         Sign(),
         MaxDeviation(),
@@ -203,7 +203,13 @@ def test_invalid_nans_true(metric, train_test_dfs):
 
 @pytest.mark.parametrize(
     "metric",
-    (MSE(missing_mode="ignore"), MAE(missing_mode="ignore"), MissingCounter()),
+    (
+        MSE(missing_mode="ignore"),
+        MAE(missing_mode="ignore"),
+        MAPE(missing_mode="ignore"),
+        SMAPE(missing_mode="ignore"),
+        MissingCounter(),
+    ),
 )
 def test_invalid_single_nan_ignore(metric, train_test_dfs):
     """Check metrics behavior in case of ignoring one nan in true values."""
@@ -221,6 +227,8 @@ def test_invalid_single_nan_ignore(metric, train_test_dfs):
     (
         (MSE(mode="per-segment", missing_mode="ignore"), type(None)),
         (MAE(mode="per-segment", missing_mode="ignore"), type(None)),
+        (MAPE(mode="per-segment", missing_mode="ignore"), type(None)),
+        (SMAPE(mode="per-segment", missing_mode="ignore"), type(None)),
         (MissingCounter(mode="per-segment"), float),
     ),
 )
@@ -243,7 +251,13 @@ def test_invalid_segment_nans_ignore_per_segment(metric, expected_type, train_te
 
 @pytest.mark.parametrize(
     "metric",
-    (MSE(mode="macro", missing_mode="ignore"), MAE(mode="macro", missing_mode="ignore"), MissingCounter(mode="macro")),
+    (
+        MSE(mode="macro", missing_mode="ignore"),
+        MAE(mode="macro", missing_mode="ignore"),
+        MAPE(mode="macro", missing_mode="ignore"),
+        SMAPE(mode="macro", missing_mode="ignore"),
+        MissingCounter(mode="macro"),
+    ),
 )
 def test_invalid_segment_nans_ignore_macro(metric, train_test_dfs):
     """Check macro metrics behavior in case of ignoring segment of all nans in true values."""
@@ -258,6 +272,8 @@ def test_invalid_segment_nans_ignore_macro(metric, train_test_dfs):
     (
         (MSE(mode="macro", missing_mode="ignore"), type(None)),
         (MAE(mode="macro", missing_mode="ignore"), type(None)),
+        (MAPE(mode="macro", missing_mode="ignore"), type(None)),
+        (SMAPE(mode="macro", missing_mode="ignore"), type(None)),
         (MissingCounter(mode="macro"), float),
     ),
 )